rename ManeuverViewHelper classes to proper name; set up the ktlint and reformat the MapboxNavigationNotification.kt
