--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/maneuver/ManeuverViewUpdate.java
@@ -1,8 +1,0 @@
-package com.mapbox.services.android.navigation.v5.internal.navigation.maneuver;
-
-import android.graphics.Canvas;
-import android.graphics.PointF;
-
-public interface ManeuverViewUpdate {
-  void updateManeuverView(Canvas canvas, int primaryColor, int secondaryColor, PointF size, float roundaboutAngle);
-}
